user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["Ammar Aamir <ammaraamir609@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.1.18"}
pytest = "^8.2.0"
httpx = "^0.27.0"
aiokafka = "^0.10.0"
protobuf = "^5.27.0"
passlib = "^1.7.4"  # Password hashing
python-jose = "^3.3.0"  # JWT creation and verification
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
